Switch to including Hacker News URLs in `readingList.csv`
This commit configures the site generator to read Hacker News URLs
for each entry in the reading list from the master CSV file that
contains all other bits of data, instead of re-querying the search
API every single time the site needs to be built.

This is also more efficient from a ratelimiting standpoint, since
only 10,000 requests to the search API are allowed in the span of
an hour.

This commit updates the struct used for (un)marshaling data into
and out of the CSV file and updates the format of the CSV file to
match this.

[skip ci]

Signed-off-by: AKP <tom@tdpain.net>